Update README.md
Typo.

This is interesting so thanks for making it available.  Open Data is good, too, thanks.

ArcMap is the opposite of that, though.  So is Elsevier. :)

I am presuming for the graphic available https://ars.els-cdn.com/content/image/1-s2.0-S0169136821003267-ga1_lrg.jpg with the abstract that the ArcMap are point location type maps from your datasets here?

Thanks,

Richard